LU-6070 libcfs: provide separate buffers for libcfs_*2str()
Provide duplicates with separate buffers for libcfs_*2str() functions.

Replace libcfs_nid2str() with libcfs_nid2str_r() function in critical
places.

Provide buffer size for nf_addr2str functions.

Use __u32 as nf_type always

+++ lnet/lnet/api-ni.c
@@ -366,8 +366,7 @@ static void lnet_assert_wire_constants(void)
         CLASSERT ((int)sizeof(((lnet_hdr_t *)0)->msg.hello.type) == 4);
 }
 
-static lnd_t *
-lnet_find_lnd_by_type (int type)
+static lnd_t *lnet_find_lnd_by_type(__u32 type)
 {
 	lnd_t		 *lnd;
 	struct list_head *tmp;
@@ -376,7 +375,7 @@ lnet_find_lnd_by_type (int type)
 	list_for_each(tmp, &the_lnet.ln_lnds) {
 		lnd = list_entry(tmp, lnd_t, lnd_list);
 
-		if ((int)lnd->lnd_type == type)
+		if (lnd->lnd_type == type)
 			return lnd;
 	}
 	return NULL;
@@ -1458,7 +1457,7 @@ lnet_startup_lndnis(struct list_head *nilist, __s32 peer_timeout,
 {
 	int			rc = 0;
 	struct lnet_ni		*ni;
-	int			lnd_type;
+	__u32			lnd_type;
 	lnd_t			*lnd;
 	struct lnet_tx_queue	*tq;
 	int			i;
